make AVs use &PL_sv_placeholder for empty slots
This is an experiment.  It is unfinished.  It turns out this is more
complicated and involves more lines of changes than using NULL for
nonexistent array elements, as there are already null checks in vari-
ous places.

This is the sort of thing that currently fails:

    my @a; $a[2] = 1; for (@a) { $_ = 2 }

because foreach needs to take &PL_sv_placeholder into account.

I’m not sure it’s worth finishing this patch.