[perl #77814] Make defelems propagate pos
[This commit was earlier reverted to make rebasing easier; its now
added back in its original form, except that the changes to pp_hot.c have
been reapplied by hand to account for the significant changes that have
taken place in pp_match() since then.]

When elements of @_ refer to nonexistent hash or array elements, then
the magic scalar in $_[0] delegates all set/get actions to the element
in represents, vivifying it if needed.

pos($_[0]), however, was not delegating the value to the element, but
storing it on the magical ‘deferred element’ scalar.