Stop minlen regexp optimisation from rejecting long strings
This fixes #112790 and part of #116907.

The length of the string is cast to I32, so it wraps and end up less
than the minimum length.

I don’t know whether the type of minlen can change without breaking
the regexp API, so simply skip this optimisation if minlen itself
wraps and becomes negative (not really testable, since the regexp com-
piler consumes far too much memory if the regexp itself is 2GB long).