regcomp.c: Debug output clearer ranges
It is confusing to have a range that isn't one of the easy to grok
all-digits, all-lowercase or all-uppercase.  It isn't very clear, for
example what all is in the range [A-z].  This commit changes the dump
(which is output when the command line option -Dr is specified, for
example) so that if a range in it includes more than just the obvious
things, it will output the range not like "A-z", but instead
"\x{41}-\x{7A}".  But things like "A-Z", "0-9" will still be output as
that.

There is another place where I didn't change to use this new technique,
later on in Perl_regprop.  This would have been more complicated to
change, and should only output such confusing things rarely, namely when
what is getting dumped includes a user-defined Unicode-like property
whose definition wasn't known at regex compile time.